+++ src/assets/navigation/check.tsx
@@ -6,9 +6,9 @@ interface Props {
 
 const CheckImg = ({ color }: Props) => {
   return (
-    <Wrapper
-      width="28"
-      height="28"
+    <svg
+      width="24"
+      height="24"
       viewBox="0 0 28 28"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
@@ -17,13 +17,8 @@ const CheckImg = ({ color }: Props) => {
         d="M12.04 16.52L8.995 13.475C8.73833 13.2183 8.42333 13.09 8.05 13.09C7.67667 13.09 7.35 13.23 7.07 13.51C6.81333 13.7667 6.685 14.0933 6.685 14.49C6.685 14.8867 6.81333 15.2133 7.07 15.47L11.06 19.46C11.3167 19.7167 11.6433 19.845 12.04 19.845C12.4367 19.845 12.7633 19.7167 13.02 19.46L20.965 11.515C21.2217 11.2583 21.35 10.9433 21.35 10.57C21.35 10.1967 21.21 9.87 20.93 9.59C20.6733 9.33333 20.3467 9.205 19.95 9.205C19.5533 9.205 19.2267 9.33333 18.97 9.59L12.04 16.52ZM14 28C12.0633 28 10.2433 27.6323 8.54 26.8968C6.83667 26.1623 5.355 25.165 4.095 23.905C2.835 22.645 1.83773 21.1633 1.1032 19.46C0.367733 17.7567 0 15.9367 0 14C0 12.0633 0.367733 10.2433 1.1032 8.54C1.83773 6.83667 2.835 5.355 4.095 4.095C5.355 2.835 6.83667 1.83727 8.54 1.1018C10.2433 0.367267 12.0633 0 14 0C15.9367 0 17.7567 0.367267 19.46 1.1018C21.1633 1.83727 22.645 2.835 23.905 4.095C25.165 5.355 26.1623 6.83667 26.8968 8.54C27.6323 10.2433 28 12.0633 28 14C28 15.9367 27.6323 17.7567 26.8968 19.46C26.1623 21.1633 25.165 22.645 23.905 23.905C22.645 25.165 21.1633 26.1623 19.46 26.8968C17.7567 27.6323 15.9367 28 14 28Z"
         fill={color ? "#F9F7FA" : "#1F1E21"}
       />
-    </Wrapper>
+    </svg>
   );
 };
-
-const Wrapper = styled.svg`
-  margin-left: 24px;
-  margin-right: 20px;
-`;
 
 export default CheckImg;
