+++ src/assets/navigation/out.tsx
@@ -6,9 +6,9 @@ interface Props {
 
 const OutImg = ({ color }: Props) => {
   return (
-    <Wrapper
-      width="40"
-      height="28"
+    <svg
+      width="24"
+      height="24"
       viewBox="0 0 40 28"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
@@ -17,13 +17,8 @@ const OutImg = ({ color }: Props) => {
         d="M32.0578 25.7522C30.543 25.7494 29.091 25.1433 28.0197 24.0666C26.9484 22.9898 26.3452 21.5302 26.3422 20.0073C26.3434 19.0197 26.598 18.0491 27.0813 17.1892C27.5646 16.3294 28.2604 15.6094 29.1014 15.0988L29.1269 15.0848L31.0135 20.3841C31.1713 20.8227 31.5809 21.1307 32.0624 21.1307C32.192 21.1297 32.3204 21.106 32.4418 21.0607L32.4337 21.063C32.5718 21.013 32.6987 20.936 32.8071 20.8366C32.9156 20.7372 33.0034 20.6173 33.0657 20.4837C33.128 20.3501 33.1635 20.2055 33.1701 20.0581C33.1768 19.9107 33.1544 19.7634 33.1044 19.6247L33.1067 19.6329L31.2201 14.3301C31.4695 14.2893 31.7584 14.2636 32.052 14.2613H32.0555C33.5703 14.2641 35.0223 14.8702 36.0936 15.9469C37.1648 17.0237 37.768 18.4833 37.7711 20.0062C37.7686 21.5288 37.1661 22.9895 36.0955 24.0664C35.0249 25.1434 33.5735 25.75 32.0589 25.7534L32.0578 25.7522ZM19.4338 17.8902L15.809 10.5846L25.3072 10.0912L19.4338 17.8902ZM15.809 18.8852C15.5131 16.8035 14.4129 14.9229 12.747 13.6512L12.7273 13.6372L13.8342 11.6239L17.4369 18.8852H15.809ZM9.8393 18.8852L11.6366 15.6191C12.6166 16.4495 13.2911 17.5864 13.5522 18.8479L13.558 18.8852H9.8393ZM7.95034 25.7522C6.43551 25.7494 4.9835 25.1433 3.91225 24.0666C2.841 22.9898 2.23779 21.5302 2.23473 20.0073C2.23749 18.4845 2.84039 17.0247 3.91143 15.9478C4.98247 14.8708 6.43435 14.2644 7.94918 14.2613C8.5711 14.2613 9.16865 14.3616 9.72791 14.5471L9.6873 14.5354L6.97569 19.4708C6.8808 19.6418 6.83254 19.835 6.8358 20.0308C6.83906 20.2267 6.89372 20.4182 6.99426 20.5859L6.99194 20.5812C7.18919 20.9137 7.54424 21.1318 7.95034 21.133H13.5522C13.0069 23.7855 10.7026 25.751 7.94222 25.7522H7.95034ZM32.0613 12.017H32.052C31.4881 12.017 30.9381 12.0765 30.4067 12.1885L30.4589 12.1792L26.8225 1.9632C26.7618 1.79542 26.6623 1.64455 26.5322 1.52306C26.4021 1.40157 26.2451 1.31294 26.0741 1.26448L26.066 1.26215L21.6105 0.0396897C21.3265 -0.0384187 21.0233 -0.000297263 20.7671 0.145719C20.511 0.291735 20.3228 0.533766 20.2437 0.818892L20.2413 0.827058C20.1643 1.11194 20.2022 1.41591 20.3469 1.67281C20.4916 1.92971 20.7314 2.11876 21.0141 2.19883L21.0222 2.20116L24.9208 3.26965L26.5336 7.78739L14.7288 8.39979L14.289 7.5156H15.577C15.7232 7.51607 15.8681 7.48745 16.0033 7.43139C16.1385 7.37534 16.2614 7.29296 16.3648 7.18899C16.4682 7.08502 16.5502 6.96152 16.6059 6.82559C16.6617 6.68966 16.6901 6.54399 16.6897 6.39696C16.6897 6.09997 16.5723 5.81397 16.3634 5.60396C16.1545 5.39396 15.8712 5.27598 15.5758 5.27598H9.17097C8.87555 5.27598 8.59223 5.39396 8.38334 5.60396C8.17444 5.81397 8.05709 6.0988 8.05709 6.39579C8.05663 6.54282 8.0851 6.68966 8.14085 6.82559C8.19661 6.96152 8.27855 7.08502 8.38197 7.18899C8.48539 7.29296 8.60824 7.37534 8.74345 7.43139C8.87865 7.48745 9.02356 7.51607 9.16981 7.5156H11.7898L12.6252 9.19532L10.7768 12.5548C9.87543 12.202 8.91645 12.0216 7.94918 12.0228C5.8428 12.0247 3.82198 12.8663 2.332 14.3631C0.842027 15.86 0.00337874 17.8898 0 20.0073C0.00214901 22.1254 0.839899 24.1574 2.32947 25.6553C3.81905 27.1533 5.83882 27.9961 7.9457 27.9988C9.85079 27.9954 11.6914 27.3046 13.1326 26.052C14.5737 24.7995 15.5198 23.0684 15.7986 21.1738L15.8032 21.1307H19.233C19.4043 21.1307 19.5733 21.0911 19.7268 21.0148C19.8804 20.9385 20.0143 20.8277 20.1184 20.6909L20.1207 20.6886L27.5848 10.7747L28.3576 12.9467C25.8142 14.3126 24.1121 16.9617 24.1086 20.0097V20.0108C24.1108 22.1288 24.9485 24.1595 26.4382 25.6572C27.9278 27.155 29.9475 27.9975 32.0543 28C34.1609 27.9975 36.1805 27.1551 37.67 25.6576C39.1596 24.1601 39.9975 22.1298 40 20.012C39.9972 17.8944 39.1592 15.8644 37.6696 14.3671C36.1801 12.8699 34.1607 12.0276 32.0543 12.0252L32.0613 12.017Z"
         fill={color ? "#F9F7FA" : "#1F1E21"}
       />
-    </Wrapper>
+    </svg>
   );
 };
-
-const Wrapper = styled.svg`
-  margin-left: 18px;
-  margin-right: 14px;
-`;
 
 export default OutImg;
