+++ src/assets/navigation/people.tsx
@@ -6,9 +6,9 @@ interface Props {
 
 const PeopleImg = ({ color }: Props) => {
   return (
-    <Wrapper
-      width="29"
-      height="21"
+    <svg
+      width="24"
+      height="24"
       viewBox="0 0 29 21"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
@@ -17,13 +17,8 @@ const PeopleImg = ({ color }: Props) => {
         d="M20.3 18.2V21H0V18.2C0 18.2 0 12.6 10.15 12.6C20.3 12.6 20.3 18.2 20.3 18.2ZM15.225 4.90006C15.225 3.93093 14.9274 2.98357 14.3697 2.17777C13.8121 1.37197 13.0195 0.743931 12.0921 0.373063C11.1648 0.00219472 10.1444 -0.0948413 9.15992 0.0942256C8.17546 0.283293 7.27119 0.749971 6.56143 1.43525C5.85168 2.12052 5.36834 2.99361 5.17252 3.94412C4.9767 4.89462 5.0772 5.87984 5.46131 6.7752C5.84543 7.67055 6.4959 8.43583 7.33048 8.97424C8.16506 9.51266 9.14626 9.80004 10.15 9.80004C11.496 9.80004 12.7868 9.28379 13.7386 8.36487C14.6903 7.44594 15.225 6.19961 15.225 4.90006ZM20.213 12.6C21.1044 13.2661 21.8338 14.1126 22.3498 15.0801C22.8659 16.0476 23.1561 17.1126 23.2 18.2V21H29V18.2C29 18.2 29 13.118 20.213 12.6ZM18.85 7.36103e-05C17.8519 -0.00529217 16.8758 0.282811 16.0515 0.826071C16.9323 2.0143 17.4059 3.43886 17.4059 4.90006C17.4059 6.36125 16.9323 7.78581 16.0515 8.97404C16.8758 9.5173 17.8519 9.8054 18.85 9.80004C20.196 9.80004 21.4868 9.28379 22.4386 8.36487C23.3903 7.44594 23.925 6.19961 23.925 4.90006C23.925 3.6005 23.3903 2.35417 22.4386 1.43525C21.4868 0.516321 20.196 7.36103e-05 18.85 7.36103e-05Z"
         fill={color ? "#F9F7FA" : "#1F1E21"}
       />
-    </Wrapper>
+    </svg>
   );
 };
-
-const Wrapper = styled.svg`
-  margin-left: 21px;
-  margin-right: 21px;
-`;
 
 export default PeopleImg;
