+++ src/assets/navigation/teacher.tsx
@@ -6,9 +6,9 @@ interface Props {
 
 const TeacherImg = ({ color }: Props) => {
   return (
-    <Wrapper
-      width="32"
-      height="27"
+    <svg
+      width="24"
+      height="24"
       viewBox="0 0 32 27"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
@@ -17,13 +17,8 @@ const TeacherImg = ({ color }: Props) => {
         d="M15.178 27C14.9231 27 14.6744 26.968 14.4317 26.9041C14.1901 26.8391 13.9546 26.7421 13.7252 26.6132L6.07885 22.4355C5.5691 22.1519 5.17403 21.7717 4.89367 21.2952C4.6133 20.8176 4.47312 20.2822 4.47312 19.6891V12.2622L0.802867 10.212C0.522501 10.0573 0.318598 9.8639 0.191159 9.6318C0.0637196 9.39971 0 9.14183 0 8.85817C0 8.5745 0.0637196 8.31662 0.191159 8.08453C0.318598 7.85244 0.522501 7.65903 0.802867 7.5043L13.7252 0.386819C13.9546 0.25788 14.1901 0.160917 14.4317 0.0959309C14.6744 0.0319768 14.9231 0 15.178 0C15.4329 0 15.6817 0.0319768 15.9243 0.0959309C16.1659 0.160917 16.4014 0.25788 16.6308 0.386819L31.1971 8.43267C31.452 8.5616 31.6498 8.74831 31.7905 8.99278C31.9302 9.23828 32 9.50287 32 9.78653V19.6891C32 20.1275 31.8532 20.4947 31.5596 20.7908C31.267 21.0879 30.904 21.2364 30.4707 21.2364C30.0374 21.2364 29.6745 21.0879 29.3819 20.7908C29.0883 20.4947 28.9415 20.1275 28.9415 19.6891V10.5602L25.8829 12.2622V19.6891C25.8829 20.2822 25.7427 20.8176 25.4624 21.2952C25.182 21.7717 24.7869 22.1519 24.2772 22.4355L16.6308 26.6132C16.4014 26.7421 16.1659 26.8391 15.9243 26.9041C15.6817 26.968 15.4329 27 15.178 27ZM15.178 14.5831L25.6535 8.85817L15.178 3.13324L4.70251 8.85817L15.178 14.5831ZM15.178 23.9054L22.8244 19.7278V13.8868L16.6691 17.3295C16.4397 17.4585 16.1975 17.5549 15.9427 17.6189C15.6878 17.6838 15.4329 17.7163 15.178 17.7163C14.9231 17.7163 14.6683 17.6838 14.4134 17.6189C14.1585 17.5549 13.9164 17.4585 13.687 17.3295L7.53166 13.8868V19.7278L15.178 23.9054Z"
         fill={color ? "#F9F7FA" : "#1F1E21"}
       />
-    </Wrapper>
+    </svg>
   );
 };
-
-const Wrapper = styled.svg`
-  margin-left: 19px;
-  margin-right: 19px;
-`;
 
 export default TeacherImg;
