--- src/common/modal/main.tsx
@@ -1,9 +1,0 @@
-import React, { useEffect, useState } from 'react';
-
-const Main=()=>{
-    const[isModal, setModal]=useState(false);
-
-    return(
-        
-    )
-}
