+++ src/components/common/modal/main.tsx
@@ -1,0 +1,7 @@
+import React, { useEffect, useState } from "react";
+
+const Main = () => {
+  const [isModal, setModal] = useState(false);
+
+  return <></>;
+};
