+++ src/components/common/modal/modal.tsx
@@ -15,9 +15,7 @@ function Modal({
       <Backdrop
         onClick={(e: React.MouseEvent) => {
           e.preventDefault();
-          if (onClickToggleModal) {
-            onClickToggleModal();
-          }
+          onClickToggleModal?.();
         }}
       />
     </ModalContainer>
